# foodcart-saas/openapi.yaml
openapi: 3.0.2
info:
  title: Foodcart SaaS API
  description: |
    Core API contract for the Foodcart SaaS platform: tenant onboarding,
    site management, content blocks, AI-assisted editing, and revisions.
  version: 0.2.0-rfc-accepted
  contact:
    name: Foodcart Engineering
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.foodcartsite.com/api/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /tenants/slug/check:
    post:
      tags: [Onboarding]
      summary: Check slug availability
      description: |
        Advises whether a proposed site slug is available. This check is
        advisory; the authoritative reservation happens during onboarding.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlugCheckRequest"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlugCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /tenants/onboard:
    post:
      tags: [Onboarding]
      summary: Create a tenant and initial site
      description: |
        Provisions a new tenant, creates the owner user from the Clerk token,
        and initializes the first site with the chosen template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantOnboardingRequest"
      responses:
        "201":
          description: Tenant and site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantOnboardingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /tenants/me:
    get:
      tags: [Tenant]
      summary: Get current tenant
      responses:
        "200":
          description: Current tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sites:
    get:
      tags: [Sites]
      summary: List sites for the current tenant
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      tags: [Sites]
      summary: Create a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteCreate"
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /sites/{site_id}:
    parameters:
      - $ref: "#/components/parameters/SiteId"
    get:
      tags: [Sites]
      summary: Get a site
      responses:
        "200":
          description: Site details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Sites]
      summary: Update a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteUpdate"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sites]
      summary: Delete a site and its content
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /sites/{site_id}/revisions:
    parameters:
      - $ref: "#/components/parameters/SiteId"
    get:
      tags: [Revisions]
      summary: List content revisions for a site
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of revisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Revision"
  /sites/{site_id}/revisions/{revision_id}/revert:
    parameters:
      - $ref: "#/components/parameters/SiteId"
      - $ref: "#/components/parameters/RevisionId"
    post:
      tags: [Revisions]
      summary: Revert site content to a previous revision
      responses:
        "200":
          description: New revision created from the revert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/sites/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9-]+$"
    get:
      tags: [Public]
      summary: Get published site content by slug
      security: []
      responses:
        "200":
          description: Public site payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSite"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/sites/{slug}/preview:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9-]+$"
      - name: preview_token
        in: query
        required: true
        schema:
          type: string
          description: Short-lived signed preview token issued by the backend.
    get:
      tags: [Public]
      summary: Preview a draft site with a signed token
      security: []
      responses:
        "200":
          description: Draft site payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSite"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT access token
  parameters:
    SiteId:
      name: site_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RevisionId:
      name: revision_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict (e.g., duplicate slug)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    SlugCheckRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9-]+(?:[a-z0-9-]+)*$"
          maxLength: 63
    SlugCheckResponse:
      type: object
      required: [slug, available]
      properties:
        slug:
          type: string
        available:
          type: boolean
        normalized_slug:
          type: string
        suggestions:
          type: array
          items:
            type: string
...
    Tenant:
      type: object
      required: [id, name, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, archived]
        billing_status:
          type: string
          enum: [trial, active, past_due, canceled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TenantOnboardingRequest:
      type: object
      required: [business_name, slug, template_id]
      properties:
        business_name:
          type: string
          maxLength: 255
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 63
        template_id:
          type: string
          enum: [banhmi, real-indian, mis-abuelos]
        initial_sources:
          $ref: "#/components/schemas/IngestionRequest"
    TenantOnboardingResponse:
      type: object
      required: [tenant, site]
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
        site:
          $ref: "#/components/schemas/Site"
    Site:
      type: object
      required: [id, tenant_id, slug, template_id, publish_state, created_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        slug:
          type: string
        template_id:
          type: string
        publish_state:
          type: string
          enum: [draft, published]
        seo:
          $ref: "#/components/schemas/SeoMeta"
        custom_domain:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteCreate:
      type: object
      required: [slug, template_id]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 63
        template_id:
          type: string
          enum: [banhmi, real-indian, mis-abuelos]
        seo:
          $ref: "#/components/schemas/SeoMeta"
    SiteUpdate:
      type: object
      properties:
        template_id:
          type: string
          enum: [banhmi, real-indian, mis-abuelos]
        publish_state:
          type: string
          enum: [draft, published]
        seo:
          $ref: "#/components/schemas/SeoMeta"
        custom_domain:
          type: string
          nullable: true
    SeoMeta:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        favicon_url:
          type: string
          format: uri
    ContentBlock:
      type: object
      required: [id, site_id, tenant_id, block_type, schema_version, data, sort_order]
      properties:
        id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        block_type:
          type: string
          enum: [hero, story, menu, locations, catering, contact, order_links, footer]
        schema_version:
          type: string
        data:
          type: object
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SocialLink:
      type: object
      required: [platform, url]
      properties:
        platform:
          type: string
          enum: [google, yelp, instagram, facebook, tiktok, website]
        url:
          type: string
          format: uri
    OrderLink:
      type: object
      required: [platform, url]
      properties:
        platform:
          type: string
          enum: [doordash, ubereats, grubhub, website, phone]
        url:
          type: string
    IngestionRequest:
      type: object
      properties:
        google_business_url:
          type: string
          format: uri
        yelp_url:
          type: string
          format: uri
        menu_url:
          type: string
          format: uri
        website_url:
          type: string
          format: uri
        social_links:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
        order_links:
          type: array
          items:
            $ref: "#/components/schemas/OrderLink"
...
    Revision:
      type: object
      required: [id, site_id, triggered_by, source, snapshot, created_at]
      properties:
        id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        triggered_by:
          type: string
          format: uuid
        source:
          type: string
          enum: [manual, ai, ingestion, revert]
        ai_request_id:
          type: string
          format: uuid
          nullable: true
        snapshot:
          type: object
          description: Full content snapshot before this revision.
        created_at:
          type: string
          format: date-time
    PublicSite:
      type: object
      required: [slug, template_id, publish_state, blocks]
      properties:
        slug:
          type: string
        template_id:
          type: string
        publish_state:
          type: string
          enum: [draft, published]
        seo:
          $ref: "#/components/schemas/SeoMeta"
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/ContentBlock"